morphology: move declineUa into Morphology so any plugin can call it
The UA declension sidecar client landed in #780 as a file-static helper in
feniaroot's root.cpp, reachable only by the Fenia NMI. Move it to
Morphology::declineUa (plug-ins/system, which fight_core and feniaroot both
already link) so the weapon generator and restring paths can decline
Ukrainian nouns/adjectives at authoring time too. The Root.declineUa NMI now
delegates to it; behaviour is unchanged (still 500ms timeout, cache,
nominative fallback).
